Refined folder structure of plans and renamed to endlessscrolllistener
