+++ app/app/src/main/java/edu/hm/foodweek/plans/screen/PlanViewPagerAdapter.kt
@@ -3,6 +3,8 @@ package edu.hm.foodweek.plans.screen
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
+import edu.hm.foodweek.plans.screen.browse.BrowsePlansFragment
+import edu.hm.foodweek.plans.screen.manage.ManagePlansFragment
 
 class PlanViewPagerAdapter(fragmentManager: FragmentManager) :
     FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
