+++ app/app/src/main/java/edu/hm/foodweek/plans/screen/browse/BrowsePlansAdapter.kt
@@ -1,4 +1,4 @@
-package edu.hm.foodweek.plans.screen
+package edu.hm.foodweek.plans.screen.browse
 
 import android.view.LayoutInflater
 import android.view.View
@@ -41,7 +41,10 @@ class BrowsePlansAdapter(
         val itemView =
             LayoutInflater.from(parent.context)
                 .inflate(R.layout.browse_plans_view_holder, parent, false)
-        return PlansViewHolder(itemView, onCardClicked)
+        return PlansViewHolder(
+            itemView,
+            onCardClicked
+        )
     }
 
     override fun getItemCount(): Int = data.size
