+++ app/app/src/main/res/layout/fragment_browse_plans.xml
@@ -13,7 +13,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".plans.screen.BrowsePlansFragment"
+        tools:context=".plans.screen.browse.BrowsePlansFragment"
         android:orientation="vertical">
 
         <androidx.appcompat.widget.SearchView
